selinux: introduce cred_selinux_ns() and use it
[PM: WARNING - DEVELOPMENT PATCH DO NOT SUBMIT UPSTREAM]

When using the SID from a cred, we should pass the selinux
namespace associated with the cred on security server calls
rather than the current selinux namespace, since they could differ.
In some of these cases, the cred is always obtained from the current
task so there is no real change, but this is cleaner and hopefully
less fragile. In other cases, the cred could in fact differ.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>
[PM: ported to v4.19-rc1, checkpatch.pl fixes]
Signed-off-by: Paul Moore <paul@paul-moore.com>