selinux: support per-namespace inode security structures
[PM: WARNING - DEVELOPMENT PATCH DO NOT SUBMIT UPSTREAM]

Extend the inode security structure to include a reference to the
associated selinux namespace, and turn it into a list so that we
can maintain per-inode security state for each namespace.  This is
necessary since the inode SIDs are per-namespace and multiple namespaces
may access the same inodes.

Introduce a find_isec() helper to find the correct inode security
structure for the current selinux namespace, creating one if one
does not already exist.  Update the existing inode_security*()
helpers to use this helper, to pass the per-namespace inode security
structure for initialization, and to return the resulting inode security
structure to the caller.  Replace direct references to inode->i_security
with the appropriate helper or to use the returned result.

This change is problematic in several respects, so it is unclear
it will survive in the final implementation.  Some of the issues are:
1) The inode security structures pin the selinux namespace in memory
for all namespaces that ever access the inode, preventing timely
(or possibly ever) freeing of the namespace.

2) Not everything in the inode security structure needs to be replicated
per namespace, so this is wasteful and potentially confusing (e.g. inode
back pointer, rcu head - but that overlaps with the list which does need
to be per-namespace, sclass unless we anticipate policy-driven file class
assignments in the future as we already have for sockets).

3) It is not safe to sleep from all callers of inode_security*(), and
thus we cannot always allocate a security blob for the namespace or
fetch the xattr on demand.  Thus, we could encounter a memory allocation
failure or fail to fetch the xattr and map it to a SID in the caller's
namespace.   This is not handled safely currently!

4) We only support a single security.selinux xattr, which must be mappable
to a SID in every namespace that accesses the inodes (or it will be mapped
to the unlabeled SID, which must then be accessible if we wish to permit
access from that namespace).

5) We do not yet properly handle setxattr of security.selinux; at present,
it will modify the on-disk xattr but will only update the in-core SID for
the current namespace and could leave other namespaces out of sync until
the inode is evicted and refetched.

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>
[PM: ported to v4.19-rc1]
Signed-off-by: Paul Moore <paul@paul-moore.com>