selinux: rename selinux state to ns (namespace)
Rename the SELinux state structures, fields, parameters,
and variables to ns since these are being used to support
SELinux namespaces.  This effectively reverts the mass renaming
that was done earlier to upstream the SELinux state encapsulation patches.

This change defines a single selinux namespace, the init_selinux_ns.
This namespace is however accessed through the current_selinux_ns
pointer in preparation for future changes where multiple selinux
namespaces will be supported.

Note that passing the current selinux namespace is not correct for
all hooks; some hooks will need to be adjusted to pass the selinux
namespace associated with an open file, a network namespace or socket,
etc, since not all hooks are invoked in process context and some
hooks operate in the context of a cred that may differ from current's
cred.  Fixing all of these cases is left to future changes, once
we introduce the support for multiple selinux namespaces.

This change by itself should have no effect on SELinux behavior or
APIs (userspace or LSM).

Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>
[PM: rebased to selinux/next on 2020-02-10]
Signed-off-by: Paul Moore <paul@paul-moore.com>