audit: cache our audit_context->sockaddr allocation
Instead of kfree()'ing the sockaddr field in the audit_context struct
at the end of every sockaddr related operation, preserve the
allocation for future use by the task.  There should be no impact on
tasks which do not do any network operations, and we'll save a
sockaddr_storage allocation for each syscall/io_uring operation beyond
the first one.

Signed-off-by: Paul Moore <paul@paul-moore.com>